* Removed symbol declaration for fntcolor, since it wasn't being defined anywhere (isn't that supposed to make the compiler cry?, am I missing something?) * Added symbol declarations for fntcolor8, fntcolor16 and fntcolor32 to libtext/libtext.h, so that other libraries can use them

+++ modules/libtext/libtext.c
@@ -52,9 +52,9 @@ extern DLVARFIXUP __bgdexport(libtext, globals_fixup)[];
 
 /* --------------------------------------------------------------------------- */
 
-int fntcolor8  = -1;
-int fntcolor16 = 0xFFFF;
-int fntcolor32 = 0xFFFFFFFF;
+int8_t fntcolor8  = -1;
+int16_t fntcolor16 = 0xFFFF;
+int32_t fntcolor32 = 0xFFFFFFFF;
 
 /* --------------------------------------------------------------------------- */
